Cleanup docker build

+++ docker-build.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-ASB_VER=3.4.5
-SRC_DIR=/root/auto-sort-bookmarks-webext-${ASB_VER}
+ASB_VER=`grep \"version\" src/manifest.json|cut -d'"' -f4`
+SRC_DIR=/root/auto-sort-bookmarks-webext
 BLD_DIR=${SRC_DIR}/build
 XPI_FILE=${SRC_DIR}/auto-sort_bookmarks-${ASB_VER}.xpi
 
@@ -15,9 +15,12 @@ docker build -t asb .
 echo
 echo Copying...
 id=$(docker create asb)
-docker cp $id:${BLD_DIR} .
-docker cp $id:${XPI_FILE} .
-docker rm -v $id
+if [ -n "$id" ]
+then
+    docker cp $id:${BLD_DIR} .
+    docker cp $id:${XPI_FILE} .
+    docker rm -v $id
+fi
 
 echo
 echo Done.
